sorrentum_sandbox/spring2024/sorrentum_sandbox/spring2023/ml_projects/Issue27_Team8_Implement_sandbox_for_Yahoo_Finance/app/requirements.txt

scikit-learn==0.24.2
dask
sqlalchemy
psycopg2-binary
yfinance
pandas
requests
tqdm
